+++ src/application/factories/user-factory.ts
@@ -5,21 +5,32 @@ import { UserEntity } from '../../domain/user/user-entity'
 class UserFactory {
   readonly ID_SIZE = 6
 
-  name: string
+  name: string = 'Tu'
+  id: string = ''
+  password: string = ''
 
-  constructor() {
-    this.name = 'Tu'
-  }
+  constructor() {}
 
   setName(name: string) {
     this.name = name
     return this
   }
 
+  setPassword(password: string) {
+    this.password = password
+    return this
+  }
+
+  setId(userId: string) {
+    this.id = userId
+    return this
+  }
+
   build(): UserEntity {
     return {
       name: this.name,
-      id: nanoid(this.ID_SIZE),
+      password: this.password,
+      id: this.id !== '' ? this.id : nanoid(this.ID_SIZE),
     }
   }
 }